requirements.txt
numpy==2.0.0
pygame==2.6.0
mistralai~=0.4.1
rich~=13.7.1
python-dotenv==1.0.1
tqdm==4.66.4
